requirements.txt
notion-client>=2.2.0
python-dotenv>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
openai>=1.57.0
